Query: Split printers to own file
